Change something in user management

+++ lib/resources/views/admin/users/index.blade.php
@@ -4,14 +4,23 @@
 
 @section('main')
 <div class="container-fluid">
-  
   <div class="row">
-    <div class="col-lg-12 margin-tb mt-5 alert alert-primary">
-        <div class="pull-left">
+      <div class="col-lg-12" style="margin-top: 40px;">
+          <ol class="breadcrumb mb-4">
+              <li class="breadcrumb-item active"><a href="{{asset('admin/')}}">Dashboard</a></li>
+              <li class="breadcrumb-item"><strong>User Management</strong></li>
+          </ol>
+
+      </div>
+  </div><!--/.row-->
+  
+  <div class="">
+    <div class="col-lg-12 margin-tb alert alert-primary">
+{{--         <div class="pull-left">
             <h2>Users Management</h2>
-        </div>
+        </div> --}}
         <div class="pull-right">
-            <a class="btn btn-success" href="{{ route('users.create') }}"> Create New User</a>
+            <a class="btn btn-success" href="{{ route('users.create') }}"><i class="fas fa-user-plus"></i> Create New User</a>
         </div>
     </div>
   
@@ -21,7 +30,7 @@
 
 
     @if ($message = Session::get('success'))
-    <div class="alert alert-success">
+    <div class="alert alert-success col-lg-12">
       <p>{{ $message }}</p>
     </div>
     @endif
@@ -52,7 +61,7 @@
         <td>
            <a class="btn btn-info" href="{{ route('users.show',$user->id) }}">Show</a>
            <a class="btn btn-primary" href="{{ route('users.edit',$user->id) }}">Edit</a>
-            {!! Form::open(['method' => 'DELETE','route' => ['users.destroy', $user->id],'style'=>'display:inline']) !!}
+            {!! Form::open(['method' => 'DELETE','route' => ['users.destroy', $user->id],'style'=>'display:inline', 'onclick' => "return confirm('are you sure?')"]) !!}
                 {!! Form::submit('Delete', ['class' => 'btn btn-danger']) !!}
             {!! Form::close() !!}
         </td>
